chore: rename plugin
