+++ packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/IndexRamBundle.ts
@@ -1,7 +1,7 @@
 import MAGIC_NUMBER from 'metro/src/shared/output/RamBundle/magic-number';
 import webpack from 'webpack';
 import { RawSource } from 'webpack-sources';
-import { Module } from './WebpackRamBundlePlugin';
+import { Module } from './RamBundleWebpackPlugin';
 
 /***
  * Reference: https://github.com/facebook/metro/blob/master/packages/metro/src/shared/output/RamBundle/as-indexed-file.js
