+++ packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/__tests__/__fixtures__/transpiled.js
