+++ packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/index.ts
@@ -1,0 +1,1 @@
+export { default } from './RamBundleWebpackPlugin';
