+++ packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/utils.ts
