+++ packages/haul-core/src/webpack/plugins/RamBundleWebpackPlugin/worker.ts
