--- packages/haul-core/src/webpack/plugins/WebpackRamBundlePlugin/index.ts
@@ -1,1 +1,0 @@
-export { default } from './WebpackRamBundlePlugin';
